refactor: make create_write_transaction public
