+++ python/deltalake/transaction.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass
+from typing import Mapping, Optional
+
+
+@dataclass
+class AddAction:
+    path: str
+    size: int
+    partition_values: Mapping[str, Optional[str]]
+    modification_time: int
+    data_change: bool
+    stats: str
